rename actors/filter_builders.py to builders.py
